karpov.courses_ml/Python/FastAPI/requirements.txt
fastapi==0.73.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
pydantic